refactor: Hooks
Refactor useSWR queries into hooks
