+++ src/js/hooks/albums.ts
@@ -1,0 +1,90 @@
+/* eslint-disable import/prefer-default-export */
+
+import useSWR, { SWRResponse } from 'swr';
+
+interface Artist {
+  external_urls: ExternalUrls;
+  href: string;
+  id: string;
+  name: string;
+  type: string;
+  uri: string;
+}
+
+interface Copyright {
+  text: string;
+  type: string;
+}
+
+interface ExternalIds {
+  upc: string;
+}
+
+interface ExternalUrls {
+  spotify: string;
+}
+
+interface Image {
+  height: number;
+  url: string;
+  width: number;
+}
+
+interface Item {
+  artists: Artist[];
+  available_markets: string[];
+  disc_number: number;
+  duration_ms: number;
+  explicit: boolean;
+  external_urls: ExternalUrls;
+  href: string;
+  id: string;
+  is_local: boolean;
+  name: string;
+  preview_url: string;
+  track_number: number;
+  type: string;
+  uri: string;
+}
+
+interface Tracks {
+  href: string;
+  items: Item[];
+  limit: number;
+  next: null;
+  offset: number;
+  previous: null;
+  total: number;
+}
+
+interface AlbumData {
+  album_type: string;
+  artists: Artist[];
+  available_markets: string[];
+  copyrights: Copyright[];
+  external_ids: ExternalIds;
+  external_urls: ExternalUrls;
+  genres: any[];
+  href: string;
+  id: string;
+  images: Image[];
+  label: string;
+  name: string;
+  popularity: number;
+  release_date: Date;
+  release_date_precision: string;
+  total_tracks: number;
+  tracks: Tracks;
+  type: string;
+  uri: string;
+}
+
+interface Error {
+  message: string;
+}
+
+function useAlbum(albumId: string): SWRResponse<AlbumData, Error> {
+  return useSWR<AlbumData, Error>(`/v1/albums/${albumId}`);
+}
+
+export { useAlbum };
