+++ src/js/hooks/playlists.ts
@@ -1,0 +1,121 @@
+/* eslint-disable import/prefer-default-export */
+
+import useSWR, { SWRResponse } from 'swr';
+
+interface Album {
+  album_type: string;
+  artists: Owner[];
+  available_markets: string[];
+  external_urls: ExternalUrls;
+  href: string;
+  id: string;
+  images: Image[];
+  name: string;
+  release_date: string;
+  release_date_precision: string;
+  total_tracks: number;
+  type: string;
+  uri: string;
+}
+
+interface ExternalIds {
+  isrc: string;
+}
+
+interface ExternalUrls {
+  spotify: string;
+}
+
+interface Followers {
+  href: null;
+  total: number;
+}
+
+interface Image {
+  height: number | null;
+  url: string;
+  width: number | null;
+}
+
+interface Item {
+  added_at: string;
+  added_by: Owner;
+  is_local: boolean;
+  primary_color: null;
+  track: Track;
+  video_thumbnail: VideoThumbnail;
+}
+
+interface Owner {
+  display_name?: string;
+  external_urls: ExternalUrls;
+  href: string;
+  id: string;
+  type: string;
+  uri: string;
+  name?: string;
+}
+
+interface Track {
+  album: Album;
+  artists: Owner[];
+  available_markets: string[];
+  disc_number: number;
+  duration_ms: number;
+  episode: boolean;
+  explicit: boolean;
+  external_ids: ExternalIds;
+  external_urls: ExternalUrls;
+  href: string;
+  id: string;
+  is_local: boolean;
+  name: string;
+  popularity: number;
+  preview_url: null | string;
+  track: boolean;
+  track_number: number;
+  type: string;
+  uri: string;
+}
+
+interface Tracks {
+  href: string;
+  items: Item[];
+  limit: number;
+  next: null;
+  offset: number;
+  previous: null;
+  total: number;
+}
+
+interface VideoThumbnail {
+  url: null;
+}
+
+interface PlaylistData {
+  collaborative: boolean;
+  description: string;
+  external_urls: ExternalUrls;
+  followers: Followers;
+  href: string;
+  id: string;
+  images: Image[];
+  name: string;
+  owner: Owner;
+  primary_color: null;
+  public: boolean;
+  snapshot_id: string;
+  tracks: Tracks;
+  type: string;
+  uri: string;
+}
+
+interface Error {
+  message: string;
+}
+
+function usePlaylist(playlistId: string): SWRResponse<PlaylistData, Error> {
+  return useSWR<PlaylistData, Error>(`/v1/playlists/${playlistId}`);
+}
+
+export { usePlaylist };
