+++ src/js/pages/Albums/Album/index.tsx
@@ -1,91 +1,14 @@
 import React from 'react';
 import { Helmet } from 'react-helmet';
 import { useParams } from 'react-router-dom';
-import useSWR from 'swr';
 import Album from '../../../components/Album';
 import Error from '../../../components/Error';
 import Loader from '../../../components/Loader';
-
-interface Artist {
-  external_urls: ExternalUrls;
-  href: string;
-  id: string;
-  name: string;
-  type: string;
-  uri: string;
-}
-
-interface Copyright {
-  text: string;
-  type: string;
-}
-
-interface ExternalIds {
-  upc: string;
-}
-
-interface ExternalUrls {
-  spotify: string;
-}
-
-interface Image {
-  height: number;
-  url: string;
-  width: number;
-}
-
-interface Item {
-  artists: Artist[];
-  available_markets: string[];
-  disc_number: number;
-  duration_ms: number;
-  explicit: boolean;
-  external_urls: ExternalUrls;
-  href: string;
-  id: string;
-  is_local: boolean;
-  name: string;
-  preview_url: string;
-  track_number: number;
-  type: string;
-  uri: string;
-}
-
-interface Tracks {
-  href: string;
-  items: Item[];
-  limit: number;
-  next: null;
-  offset: number;
-  previous: null;
-  total: number;
-}
-
-interface AlbumData {
-  album_type: string;
-  artists: Artist[];
-  available_markets: string[];
-  copyrights: Copyright[];
-  external_ids: ExternalIds;
-  external_urls: ExternalUrls;
-  genres: any[];
-  href: string;
-  id: string;
-  images: Image[];
-  label: string;
-  name: string;
-  popularity: number;
-  release_date: Date;
-  release_date_precision: string;
-  total_tracks: number;
-  tracks: Tracks;
-  type: string;
-  uri: string;
-}
+import { useAlbum } from '../../../hooks/albums';
 
 function AlbumPage(): JSX.Element {
   const { albumId } = useParams();
-  const { data, error } = useSWR<AlbumData>(`/v1/albums/${albumId}`);
+  const { data, error } = useAlbum(albumId);
 
   if (error) {
     return <Error>{error.message}</Error>;
