+++ src/js/pages/Artists/Artist/Albums/index.tsx
@@ -1,60 +1,14 @@
 import React from 'react';
 import { Helmet } from 'react-helmet';
 import { useParams } from 'react-router-dom';
-import useSWR from 'swr';
 import Albums from '../../../../components/Albums';
 import Error from '../../../../components/Error';
 import Loader from '../../../../components/Loader';
-
-interface Artist {
-  external_urls: ExternalUrls;
-  href: string;
-  id: string;
-  name: string;
-  type: string;
-  uri: string;
-}
-
-interface ExternalUrls {
-  spotify: string;
-}
-
-interface Image {
-  height: number;
-  url: string;
-  width: number;
-}
-
-interface Item {
-  album_group: string;
-  album_type: string;
-  artists: Artist[];
-  available_markets: string[];
-  external_urls: ExternalUrls;
-  href: string;
-  id: string;
-  images: Image[];
-  name: string;
-  release_date: string;
-  release_date_precision: string;
-  total_tracks: number;
-  type: string;
-  uri: string;
-}
-
-interface AlbumsData {
-  href: string;
-  items: Item[];
-  limit: number;
-  next: string;
-  offset: number;
-  previous: null;
-  total: number;
-}
+import { useArtistAlbums } from '../../../../hooks/artists';
 
 function AlbumsPage(): JSX.Element {
   const { artistId } = useParams();
-  const { data, error } = useSWR<AlbumsData>(`/v1/artists/${artistId}/albums`);
+  const { data, error } = useArtistAlbums(artistId);
 
   if (error) {
     return <Error>{error.message}</Error>;
