+++ src/js/pages/Artists/Artist/TopTracks/index.tsx
@@ -1,76 +1,14 @@
 import React from 'react';
 import { Helmet } from 'react-helmet';
 import { useParams } from 'react-router-dom';
-import useSWR from 'swr';
 import AlbumTracks from '../../../../components/AlbumTracks';
 import Error from '../../../../components/Error';
 import Loader from '../../../../components/Loader';
-
-interface Album {
-  album_type: string;
-  artists: Artist[];
-  external_urls: ExternalUrls;
-  href: string;
-  id: string;
-  images: Image[];
-  name: string;
-  release_date: string;
-  release_date_precision: string;
-  total_tracks: number;
-  type: string;
-  uri: string;
-}
-
-interface Artist {
-  external_urls: ExternalUrls;
-  href: string;
-  id: string;
-  name: string;
-  type: string;
-  uri: string;
-}
-
-interface ExternalUrls {
-  spotify: string;
-}
-
-interface Image {
-  height: number;
-  url: string;
-  width: number;
-}
-
-interface ExternalIds {
-  isrc: string;
-}
-
-interface Track {
-  album: Album;
-  artists: Artist[];
-  disc_number: number;
-  duration_ms: number;
-  explicit: boolean;
-  external_ids: ExternalIds;
-  external_urls: ExternalUrls;
-  href: string;
-  id: string;
-  is_local: boolean;
-  is_playable: boolean;
-  name: string;
-  popularity: number;
-  preview_url: string;
-  track_number: number;
-  type: string;
-  uri: string;
-}
-
-interface TopTracksData {
-  tracks: Track[];
-}
+import { useArtistTopTracks } from '../../../../hooks/artists';
 
 function TopTracksPage(): JSX.Element {
   const { artistId } = useParams();
-  const { data, error } = useSWR<TopTracksData>(`/v1/artists/${artistId}/top-tracks?market=DE`);
+  const { data, error } = useArtistTopTracks(artistId);
 
   if (error) {
     return <Error>{error.message}</Error>;
