+++ src/js/pages/Users/User/Playlists/index.tsx
@@ -1,65 +1,14 @@
 import React from 'react';
 import { Helmet } from 'react-helmet';
 import { useParams } from 'react-router-dom';
-import useSWR from 'swr';
 import Error from '../../../../components/Error';
 import Loader from '../../../../components/Loader';
 import Playlists from '../../../../components/Playlists';
-
-interface ExternalUrls {
-  spotify: string;
-}
-
-interface Image {
-  height: number;
-  url: string;
-  width: number;
-}
-
-interface Item {
-  collaborative: boolean;
-  description: string;
-  external_urls: ExternalUrls;
-  href: string;
-  id: string;
-  images: Image[];
-  name: string;
-  owner: Owner;
-  primary_color: null;
-  public: boolean;
-  snapshot_id: string;
-  tracks: Tracks;
-  type: string;
-  uri: string;
-}
-
-interface Owner {
-  display_name: string;
-  external_urls: ExternalUrls;
-  href: string;
-  id: string;
-  type: string;
-  uri: string;
-}
-
-interface Tracks {
-  href: string;
-  total: number;
-}
-
-interface PlaylistsData {
-  href: string;
-  items: Item[];
-  limit: number;
-  next: string;
-  offset: number;
-  previous: null;
-  total: number;
-}
+import { useUserPlaylists } from '../../../../hooks/users';
 
 function PlaylistsPage(): JSX.Element {
   const { userId } = useParams();
-  const { data, error } = useSWR<PlaylistsData>(`/v1/users/${userId}/playlists`);
+  const { data, error } = useUserPlaylists(userId);
 
   if (error) {
     return <Error>{error.message}</Error>;
