+++ src/js/pages/Users/User/index.tsx
@@ -1,40 +1,14 @@
 import React from 'react';
 import { Helmet } from 'react-helmet';
 import { useParams } from 'react-router-dom';
-import useSWR from 'swr';
 import Error from '../../../components/Error';
 import Loader from '../../../components/Loader';
 import User from '../../../components/User';
-
-interface ExternalUrls {
-  spotify: string;
-}
-
-interface Followers {
-  href: null;
-  total: number;
-}
-
-interface Image {
-  height: number;
-  url: string;
-  width: number;
-}
-
-interface UserData {
-  display_name: string;
-  external_urls: ExternalUrls;
-  followers: Followers;
-  href: string;
-  id: string;
-  images: Image[];
-  type: string;
-  uri: string;
-}
+import { useUser } from '../../../hooks/users';
 
 function UserPage(): JSX.Element {
   const { userId } = useParams();
-  const { data, error } = useSWR<UserData>(`/v1/users/${userId}`);
+  const { data, error } = useUser(userId);
 
   if (error) {
     return <Error>{error.message}</Error>;
